+++ src/Cmd/Repoquery.hs
@@ -4,9 +4,6 @@ where
 
 import Branches
 import Common.System
-import Repoquery
-
-import qualified Data.List as L
 
 repoqueryCmd :: (BranchesReq, [String]) -> IO ()
 repoqueryCmd (breq, pkgs) = do
@@ -14,11 +11,4 @@ repoqueryCmd (breq, pkgs) = do
            then pure <$> getDirectoryName
            else return pkgs
   brs <- listOfBranches True False breq
-  sysbr <- systemBranch
-  mapM_ (repoquery_ sysbr query) brs
-  where
-    repoquery_ :: Branch -> [String] -> Branch -> IO ()
-    repoquery_ sysbr query br = do
-      let qf = ["--queryformat=%{repoid}: %{name}-%{version}-%{release}.%{arch}"
-               | not (any ("--qf" `L.isPrefixOf`) query)]
-      repoquery sysbr br (qf ++ query) >>= putStrLn
+  cmd_ "fedora-repoquery" $ ["--qf", "nvr"] ++ map showBranch brs ++ query